add vote assertiveness preferences
Vote Assertiveness determines how pushy you want the vote system to be
with you when a vote has been called.

"Start Notify" is equivalent to prior behavior - voters will be notified
once at the start of the vote and not reminded.

"Start Popup" will do the same as start notify, but also open the vote
window automatically.

"Reminder Notify" will notify the voter a second time *if* half the vote
time has passed and they have not voted.

"Reminder Popup" will do the same as reminder notify, but also open the
vote window automatically.

This new preference defaults to Reminder Notify.

----

Also, explains to inelligible voters why they can't vote if the vote
in question will not allow it. Implements this for transfer votes.